pyproject.toml
[project]
name = "blogger-export"
version = "0.1.0"
description = ""
authors = [
    {name = "Kevin McNamee",email = "46104536+mrkmcnamee@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client~=2.172.0",
    "google-auth~=2.40.0",
    "google_auth_oauthlib~=1.2.0",
]
